Thread Subscription test done
changed all id's and foreign key id's to unsignedBigInteger from
unsignedInteger and made the user email verimication email work

+++ database/migrations/2019_03_05_170922_create_threads_table.php
@@ -15,8 +15,8 @@ class CreateThreadsTable extends Migration
     {
         Schema::create('threads', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('channel_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('channel_id');
             $table->unsignedInteger('replies_count')->default(0);
             $table->string('title');
             $table->text('body');
