+++ database/migrations/2019_03_09_091009_create_activities_table.php
@@ -15,8 +15,8 @@ class CreateActivitiesTable extends Migration
     {
         Schema::create('activities', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id')->index();
-            $table->unsignedInteger('subject_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
+            $table->unsignedBigInteger('subject_id')->index();
             $table->string('subject_type', 50);
             $table->string('type', 50);
             $table->timestamps();
